refactor: split i18n page translations
