--- src/i18n/locales/en/pages.ts
@@ -1,639 +1,0 @@
-export const pages = {
-  // App-level pages
-  app: {
-    beatmapsComingSoon: 'Beatmaps (coming soon)',
-    notFound: '404 - Page not found',
-  },
-  
-  // Home page
-  hero: {
-    tagline: 'Probably the best osu! private server',
-    description:
-      'g0v0! is a brand-new osu! private server <bold>built for the lazer client</bold>. We support the standard / taiko / catch / mania rulesets, RX/AP pp calculation, and unlimited renames so you can play your way.',
-    statusOperational: 'All services are operational',
-    statusCommunity: 'Join our QQ / Discord community for support and updates',
-    joinCta: 'How to join the server',
-    viewProfile: 'View profile',
-    viewRankings: 'View rankings',
-    register: 'Sign up',
-    login: 'Sign in',
-    featuresTitle: 'Feature highlights',
-    featuresSubtitle: 'Discover the rich features and services we provide for you',
-    community: {
-      qq: 'QQ Group',
-      discord: 'Discord',
-      github: 'GitHub',
-      discordTag: 'g0v0!',
-    },
-  },
-  
-  features: {
-    items: [
-      {
-        title: 'Works on every platform',
-        content:
-          'g0v0! is designed for lazer, which means you can play on any system that supports osu! lazer without platform limitations.',
-        imageAlt: 'Works on every platform',
-      },
-      {
-        title: 'Faster beatmap downloads',
-        content:
-          'We use Sayobot mirror services to deliver speedy downloads for players in China, so grabbing maps is no longer a waiting game.',
-        imageAlt: 'Faster beatmap downloads',
-      },
-      {
-        title: 'Active community support',
-        content:
-          'Our official QQ group and Discord server provide the perfect place to ask for help, share highlights, or just hang out.',
-        imageAlt: 'Active community support',
-      },
-      {
-        title: 'Ultimate customization',
-        content:
-          'Freely change your username, upload custom banners, and create a dazzling profile and signature—let your creativity run wild.',
-        imageAlt: 'Ultimate customization',
-      },
-      {
-        title: 'Developer friendly',
-        content:
-          'Our server follows the official osu! v1/v2 API specs so you can integrate bots and services into g0v0! faster and easier.',
-        imageAlt: 'Developer friendly',
-      },
-      {
-        title: 'Open and transparent',
-        content:
-          'Both the server and client are open source. The server also provides hook support so you can extend it with plugins (in development).',
-        imageAlt: 'Open and transparent',
-      },
-      {
-        title: 'User-submitted beatmaps',
-        content:
-          'Submit your own beatmaps or ranked-ineligible maps from the official server and get them approved here (in development).',
-        imageAlt: 'User-submitted beatmaps',
-      },
-      {
-        title: 'Ruleset leaderboards',
-        content:
-          'We support custom ruleset score calculations and submissions, plus robust leaderboard support (in development).',
-        imageAlt: 'Ruleset leaderboards',
-      },
-    ],
-  },
-  
-  // Rankings page
-  rankings: {
-    title: 'Rankings',
-    tabs: {
-      users: 'User Rankings',
-      countries: 'Country Rankings',
-    },
-    rankingTypes: {
-      performance: 'Performance',
-      score: 'Score',
-    },
-    gameModes: {
-      osu: 'osu!',
-      taiko: 'osu!taiko',
-      fruits: 'osu!catch',
-      mania: 'osu!mania',
-    },
-    subModes: {
-      vanilla: 'Vanilla',
-      relax: 'Relax',
-      autopilot: 'Autopilot',
-    },
-    filters: {
-      country: 'Country',
-      allCountries: 'All Countries',
-    },
-    userCard: {
-      rank: 'Rank',
-      performance: 'Performance',
-      accuracy: 'Accuracy',
-      playCount: 'Play Count',
-      country: 'Country',
-    },
-    countryCard: {
-      rank: 'Rank',
-      country: 'Country',
-      averagePerformance: 'Average Performance',
-      totalPerformance: 'Total Performance',
-      activeUsers: 'active users',
-      playCount: 'plays',
-    },
-    pagination: {
-      previous: 'Previous',
-      next: 'Next',
-      page: 'Page {{current}} of {{total}}',
-    },
-    errors: {
-      loadFailed: 'Failed to load rankings',
-      noData: 'No data available',
-    },
-  },
-  
-  // Settings page
-  settings: {
-    title: 'Account Settings',
-    description: 'Manage your account information and preferences',
-    errors: {
-      loadFailed: 'Unable to load settings',
-      tryRefresh: 'Please try refreshing the page'
-    },
-    username: {
-      title: 'Username Settings',
-      current: 'Current Username',
-      change: 'Change Username',
-      placeholder: 'Enter new username',
-      hint: 'After changing your username, your old username will be saved in history',
-      save: 'Save',
-      saving: 'Saving...',
-      cancel: 'Cancel',
-      success: 'Username changed successfully!',
-      errors: {
-        empty: 'Username cannot be empty',
-        sameAsOld: 'New username is the same as current username',
-        taken: 'Username is taken, please choose another one',
-        userNotFound: 'User not found',
-        failed: 'Failed to change username, please try again later'
-      }
-    },
-    avatar: {
-      title: 'Avatar Settings',
-      current: 'Current Avatar',
-      change: 'Change Avatar',
-      hint: 'Supports PNG, JPEG, GIF formats, recommended size 256x256 pixels, max 5MB',
-      success: 'Avatar changed successfully!'
-    },
-    cover: {
-      title: 'Cover Settings',
-      label: 'Profile Cover Image',
-      hint: 'Recommended size: 2000x500 pixels (4:1 ratio recommended), supports PNG, JPEG, GIF formats, max 10MB'
-    },
-    account: {
-      title: 'Account Information',
-      userId: 'User ID',
-      joinDate: 'Join Date',
-      country: 'Country/Region',
-      lastVisit: 'Last Visit'
-    }
-  },
-  
-  // Teams page
-  teams: {
-    title: 'Team Rankings',
-    description: 'View the performance and ranking of teams',
-    tabs: {
-      teams: 'Team Rankings',
-    },
-    createTeam: 'Create Team',
-    myTeam: 'My Team',
-    joinTeam: 'Join Team',
-    viewTeam: 'View Team',
-    editTeam: 'Edit Team',
-    teamRankings: 'Team Rankings',
-    loadingTeams: 'Loading team rankings data...',
-    gameModes: {
-      osu: 'osu!',
-      taiko: 'osu!taiko',
-      fruits: 'osu!catch',
-      mania: 'osu!mania',
-    },
-    rankingTypes: {
-      performance: 'Performance',
-      score: 'Score',
-    },
-    rankingType: {
-      label: 'Ranking Type',
-    },
-    subModes: {
-      vanilla: 'Vanilla',
-      relax: 'Relax',
-      autopilot: 'Autopilot',
-    },
-    teamCard: {
-      rank: 'Rank',
-      members: 'Members',
-      captain: 'Captain',
-      averageRank: 'Average Rank',
-      totalPerformance: 'Total Performance',
-      country: 'Country',
-    },
-    pagination: {
-      previous: 'Previous',
-      next: 'Next',
-      page: 'Page {{current}} of {{total}}',
-    },
-    errors: {
-      loadFailed: 'Failed to load teams',
-      noData: 'No team data available',
-    },
-    detail: {
-      createdAt: 'Created on {{date}}',
-      members: '{{count}} members',
-      captain: 'Captain',
-      membersSectionTitle: 'Members',
-      teamMembers: 'Team Members',
-      memberCount: '({{count}} people)',
-      loading: 'Loading team details...',
-      notFound: 'Team not found',
-      notFoundDescription: 'The team information was not found',
-      backToTeams: 'Back to teams',
-      teamInfo: 'Team header information',
-      coverImage: 'Cover image',
-      teamFlag: 'Team flag',
-      teamBasicInfo: 'Team basic information',
-      teamActions: 'Team action buttons',
-      captainInfo: 'Captain information',
-      memberActions: 'Member action buttons',
-      joinTeam: 'Request to Join',
-      joining: 'Requesting...',
-      leaveTeam: 'Leave Team',
-      leaving: 'Leaving...',
-      editTeam: 'Edit Team',
-      deleteTeam: 'Delete Team',
-      kickMember: 'Kick',
-      kicking: 'Kicking...',
-      joinRequestSent: 'Join request sent, please wait for captain approval',
-      leftTeam: 'Left the team',
-      confirmLeave: 'Are you sure you want to leave this team?',
-      confirmDelete: 'Are you sure you want to delete this team? This action cannot be undone!',
-      confirmKick: 'Are you sure you want to kick {{username}}?',
-      teamDeleted: 'Team deleted',
-      memberKicked: '{{username}} has been kicked from the team',
-    },
-    create: {
-      loginRequired: 'You need to login to create or edit teams',
-      nameRequired: 'Please enter team name',
-      shortNameRequired: 'Please enter team short name',
-      assetsRequired: 'Please upload team flag and cover',
-      updateSuccess: 'Team information updated successfully',
-      createSuccess: 'Team created successfully',
-      loading: 'Loading team information...',
-      editTeam: 'Edit Team',
-      createTeam: 'Create Team',
-      editDescription: 'Modify your team information',
-      createDescription: 'Create a new team and invite friends to play together',
-      basicInfo: 'Basic Information',
-      teamName: 'Team Name',
-      teamNamePlaceholder: 'Enter team name',
-      teamShortName: 'Team Short Name',
-      teamShortNamePlaceholder: 'Enter team short name',
-      shortNameDescription: 'Short name will be displayed in rankings, recommend using 2-5 characters',
-      teamFlag: 'Team Flag',
-      selectFlag: 'Select flag',
-      flagDescription: 'Standard size: 240×120px, max 2MB, supports cropping',
-      teamCover: 'Team Cover',
-      selectCover: 'Select cover',
-      coverDescription: 'Recommended size: 1920×1280px, max 10MB, supports cropping',
-      creatingTeam: 'Creating team...',
-      memberManagement: 'Member Management',
-      leaderTransfer: 'Transfer Leadership (Optional)',
-      keepCurrentLeader: 'Keep current leader',
-      leaderTransferDescription: 'Select a new leader. If not selected, current leader will remain. After transferring leadership, you will lose management permissions.',
-      currentMembers: 'Current members ({{count}} people)',
-      currentLeader: 'Current leader',
-      willBecomeLeader: 'Will become new leader',
-      noMembersAvailable: 'No members available for selection',
-      unknown: 'Unknown',
-      cancel: 'Cancel',
-      saving: 'Saving...',
-      saveChanges: 'Save Changes',
-      required: '*',
-    },
-  },
-  
-  // Messages page
-  messages: {
-    title: 'Messages',
-    tabs: {
-      channels: 'Channels',
-      notifications: 'Notifications',
-    },
-    channels: {
-      general: 'General',
-      osu: 'osu!',
-      taiko: 'osu!taiko',
-      fruits: 'osu!catch',
-      mania: 'osu!mania',
-    },
-    chat: {
-      placeholder: 'Type a message...',
-      send: 'Send',
-      online: 'Online',
-      offline: 'Offline',
-      typing: 'Typing...',
-    },
-    notifications: {
-      markAllRead: 'Mark all as read',
-      refresh: 'Refresh',
-      empty: 'No notifications',
-      types: {
-        friend_request: 'Friend Request',
-        team_invite: 'Team Invitation',
-        system: 'System Notification',
-      },
-    },
-    privateMessage: {
-      title: 'Private Message',
-      newMessage: 'New Message',
-      sendTo: 'Send to',
-      close: 'Close',
-    },
-    errors: {
-      loadFailed: 'Failed to load messages',
-      sendFailed: 'Failed to send message',
-      connectFailed: 'Connection failed',
-    },
-  },
-  
-  // Profile/User pages
-  profile: {
-    errors: {
-      loadFailed: 'Unable to load profile',
-      userNotFound: 'User not found',
-      checkId: 'Please check if the user ID or username is correct',
-      tryRefresh: 'Please try refreshing the page',
-    },
-    info: {
-      title: 'Player Information',
-      globalRank: 'Global Rank',
-      countryRank: 'Country Rank',
-      underConstruction: 'Remaining data is under construction',
-    },
-    stats: {
-      rankedScore: 'Ranked Score',
-      accuracy: 'Accuracy',
-      playCount: 'Play Count',
-      totalScore: 'Total Score',
-      totalHits: 'Total Hits',
-      hitsPerPlay: 'Hits Per Play',
-      maxCombo: 'Max Combo',
-      replaysWatched: 'Replays Watched',
-      medals: 'Medals',
-      pp: 'PP',
-      playTime: 'Play Time',
-    },
-    activities: {
-      title: 'Recent Activity',
-      noActivities: 'No recent activity',
-      loadFailed: 'Failed to load user activities',
-      loadMore: 'Load More',
-      loading: 'Loading...',
-      timeAgo: {
-        justNow: 'Just now',
-        minutesAgo: '{{count}} minutes ago',
-        hoursAgo: '{{count}} hours ago',
-        daysAgo: '{{count}} days ago',
-        monthsAgo: '{{count}} months ago',
-        yearsAgo: '{{count}} years ago',
-      },
-      types: {
-        rank: {
-          prefix: 'on',
-          middle: 'achieved',
-          grade: 'rank',
-          rankPrefix: 'ranked #',
-        },
-        rankLost: {
-          prefix: 'on',
-          suffix: 'lost first place',
-        },
-        achievement: 'unlocked achievement',
-        beatmapUpload: 'uploaded beatmap',
-        beatmapUpdate: 'updated beatmap',
-        beatmapRanked: 'beatmap was ranked',
-        beatmapLoved: 'beatmap was loved',
-        beatmapRevived: 'beatmap was revived',
-        beatmapDeleted: 'beatmap was deleted',
-        usernameChange: 'changed username',
-        supportAgain: 'became a Supporter again',
-        supportFirst: 'became a Supporter for the first time',
-        supportGift: 'received Supporter as a gift',
-        activity: 'performed an activity',
-      },
-    },
-    bestScores: {
-      title: 'Best Scores',
-      noScores: 'No best scores yet',
-      loadFailed: 'Failed to load best scores',
-      loadMore: 'Load More',
-      loading: 'Loading...',
-      by: 'by',
-    },
-    friends: {
-      loading: 'Loading...',
-    },
-    userPage: {
-      title: 'Personal Introduction',
-      noContent: 'No personal introduction written yet',
-      writeButton: 'Write Personal Introduction',
-      editButton: 'Edit',
-      editTitle: 'Edit Personal Introduction',
-      placeholder: 'Write your personal introduction using BBCode...',
-      editorPlaceholder: 'Write a personal introduction for {{username}}...\n\nYou can use BBCode to format text, such as:\n[b]Bold text[/b]\n[i]Italic text[/i]\n[color=red]Colored text[/color]\n\nClick toolbar buttons or use shortcuts to quickly insert formatting.',
-      processing: 'Content is being processed...',
-      cancel: 'Cancel',
-      save: 'Save',
-      saving: 'Saving...',
-      saveSuccess: 'Personal page saved successfully!',
-      saveError: 'Save failed, please try again',
-      hasUnsavedChanges: 'You have unsaved changes',
-      confirmDiscard: 'You have unsaved changes, are you sure you want to discard editing?',
-      noEditPermission: 'You do not have permission to edit this user\'s page',
-      cannotEditPage: 'Cannot edit this page',
-      loadingEditor: 'Loading editor...',
-      editPageTitle: 'Edit Personal Page',
-      editPageSubtitle: 'Edit personal introduction for {{username}}',
-      unsavedChanges: 'Unsaved changes',
-      cancelEditTooltip: 'Cancel editing (Esc)',
-      saveShortcut: 'Ctrl+S Save',
-      supportsBBCode: 'Supports BBCode format',
-      bbcodeDescription: 'BBCode is a lightweight markup language used to format text. You can use the following tags to beautify your personal introduction:',
-      usePreview: 'Use preview function to see the final effect',
-      editCover: 'Edit cover',
-    },
-    friendActions: {
-      selfProfile: 'Own profile',
-      blocked: 'User blocked',
-      mutualFollow: 'Mutual follow - you follow each other',
-      following: 'Following - you follow this user',
-      followsYou: 'Follows you - this user follows you',
-      notFollowing: 'Not following - click to see follow options',
-      unblock: 'Unblock',
-      cancelMutualFollow: 'Cancel mutual follow',
-      unfollow: 'Unfollow',
-      block: 'Block',
-      followBack: 'Follow back (mutual follow)',
-      follow: 'Follow',
-    },
-    bbcodeEditor: {
-      placeholder: 'Enter BBCode content here...',
-      toolbar: {
-        bold: 'Bold',
-        italic: 'Italic',
-        underline: 'Underline',
-        strikethrough: 'Strikethrough',
-        color: 'Color',
-        fontSize: 'Font Size',
-        image: 'Insert Image',
-        link: 'Insert Link',
-        userProfile: 'User Profile Link',
-        email: 'Email Link',
-        youtube: 'YouTube Video',
-        audio: 'Audio',
-        imagemap: 'Image Map',
-        quote: 'Quote',
-        code: 'Code Block',
-        list: 'List',
-        box: 'Collapsible Box',
-        spoiler: 'Spoiler',
-        center: 'Center Align',
-        heading: 'Heading',
-        notice: 'Notice Box',
-        help: 'Help',
-      },
-      insertText: {
-        bold: 'Bold text',
-        italic: 'Italic text',
-        underline: 'Underlined text',
-        strikethrough: 'Strikethrough text',
-        colorText: 'Colored text',
-        text: 'text',
-        linkText: 'Link text',
-        username: 'Username',
-        emailLink: 'Email link',
-        clickToVisit: 'Click to visit website',
-        infoArea: 'This is info area',
-        quoteContent: 'Quote content',
-        codeContent: 'Code content',
-        item1: 'Item 1',
-        item2: 'Item 2',
-        title: 'Title',
-        collapsibleContent: 'Collapsible content',
-        spoilerContent: 'Spoiler content',
-        centerText: 'Centered text',
-        headingText: 'Heading text',
-        importantNotice: 'Important notice',
-      },
-      colors: {
-        text: 'color text',
-      },
-      fontSize: {
-        label: 'Font Size',
-        extraSmall: 'Extra Small',
-        small: 'Small',
-        normal: 'Normal',
-        large: 'Large',
-      },
-      help: {
-        title: 'BBCode Help',
-        button: 'Help',
-        examples: {
-          bold: 'Bold',
-          italic: 'Italic',
-          underline: 'Underline',
-          strikethrough: 'Strikethrough',
-          color: 'Color',
-          size: 'Size',
-          link: 'link',
-          text: 'text',
-          imageUrl: 'Image URL',
-          quote: 'Quote',
-          code: 'Code',
-          item1: 'Item 1',
-          item2: 'Item 2',
-        },
-      },
-      modes: {
-        edit: 'Edit',
-        preview: 'Preview',
-      },
-      preview: {
-        noContent: 'Enter content to see preview',
-        generateFailed: 'Unable to generate preview',
-        generating: 'Generating preview...',
-      },
-      validation: {
-        networkError: 'Validation failed, please check network connection',
-        validating: 'Validating',
-        syntaxCorrect: 'Syntax correct',
-        errors: ' errors',
-        syntaxErrors: 'BBCode syntax errors',
-      },
-    },
-  },
-  
-  
-  "howToJoin": {
-    "title": "How to Join the Server",
-    "subtitle": "There are two ways to connect to our server",
-    "copyFailed": "Copy failed:",
-    "clickToCopy": "Click to copy",
-    "method1": {
-      "title": "Using Our Custom Client",
-      "recommended": "Recommended",
-      "description": "This method is recommended for all users who can run osu!lazer on their platform.",
-      "steps": {
-        "title": "Steps:",
-        "step1": {
-          "title": "Download g0v0! Custom Client",
-          "pcVersion": "PC Version:",
-          "androidVersion": "Android Version:",
-          "downloadPc": "Download from GitHub Releases",
-          "downloadAndroidDomestic": "Download (China Only)",
-          "downloadAndroidOverseas": "Download (For global network)"
-        },
-        "step2": {
-          "title": "Launch the game, go to Settings → Online, and enter in the \"Custom API Server URL\" field:",
-          "description": "In the osu!lazer settings, go to the “Online” section, find the “Custom API Server URL” setting, and enter the following text in the input box:",
-          "imageHint": "As shown in the image"
-        },
-        "step3": {
-          "title": "Restart the game and enjoy!",
-          "description": "After entering the URL, exit osu!lazer and restart it for the changes to take effect."
-        }
-      }
-    },
-    "method2": {
-      "title": "Using LazerAuthlibInjection (x86_64 Platforms Only)",
-      "suitableFor": "This method is suitable for the following users:",
-      "platforms": {
-        "windows": "Windows users. (WOA Temporarily Not Supported)",
-        "linux": "Any Linux Distributions. (arm64 and any other arm devices not supported)",
-        "mac": "macOS. (Apple Silicon are not supported)"
-      },
-      "steps": {
-        "title": "Steps:",
-        "step1": {
-          "title": "Download LazerAuthlibInjection",
-          "download": "Download from GitHub Releases",
-          "button": "Download LazerAuthlibInjection"
-        },
-        "step2": {
-          "title": "Install it as a ruleset into osu!lazer",
-          "description": "Install the downloaded LazerAuthlibInjection as a ruleset into osu!lazer"
-        },
-        "step3": {
-          "title": "Launch the game, go to Settings → Game Mode, and enter the following information:",
-          "description": "Configure the server connection info in the game settings",
-          "apiUrl": "API URL:",
-          "websiteUrl": "Website URL:"
-        },
-        "step4": {
-          "title": "After seeing the 'API Settings Changed' notification, restart the client and enjoy!",
-          "description": "After completing the setup, restart the client to connect to the server"
-        }
-      },
-      "warning": {
-        "title": "Important Notice",
-        "description": "Although peppy(ppy) explicitly states that rulesets are not subject to anti-cheat detection, we strongly recommend that you avoid connecting the osu!lazer client with AuthLibInject installed to the official servers, as doing so may result in your account being banned!"
-      }
-    }
-  },
-
-  
-  footer: {
-    copyright: '© 2025 g0v0.top. Crafted for rhythm game players.',
-  },
-} as const;
