+++ src/i18n/locales/en/pages/app.ts
@@ -1,0 +1,6 @@
+export const appPages = {
+  app: {
+    beatmapsComingSoon: 'Beatmaps (coming soon)',
+    notFound: '404 - Page not found',
+  },
+} as const;
