+++ src/i18n/locales/en/pages/layout.ts
@@ -1,0 +1,5 @@
+export const layout = {
+  footer: {
+    copyright: '© 2025 g0v0.top. Crafted for rhythm game players.',
+  },
+} as const;
