+++ src/i18n/locales/zh/pages/app.ts
@@ -1,0 +1,6 @@
+export const appPages = {
+  app: {
+    beatmapsComingSoon: '谱面（即将推出）',
+    notFound: '404 - 页面未找到',
+  },
+} as const;
