+++ src/i18n/locales/zh/pages/profile.ts
@@ -1,0 +1,217 @@
+export const profilePage = {
+  profile: {
+    errors: {
+      loadFailed: '无法加载个人资料',
+      userNotFound: '用户未找到',
+      checkId: '请检查用户 ID 或用户名是否正确',
+      tryRefresh: '请尝试刷新页面',
+    },
+    info: {
+      title: '玩家信息',
+      globalRank: '全球排名',
+      countryRank: '地区排名',
+      underConstruction: '剩下数据正在努力施工中',
+    },
+    stats: {
+      rankedScore: '计分成绩总分',
+      accuracy: '准确率',
+      playCount: '游戏次数',
+      totalScore: '总分',
+      totalHits: '总命中次数',
+      hitsPerPlay: '每次游玩击打数',
+      maxCombo: '最大连击',
+      replaysWatched: '回放被观看次数',
+      medals: '奖章',
+      pp: 'PP',
+      playTime: '游玩时间',
+    },
+    activities: {
+      title: '最近活动',
+      noActivities: '暂无最近活动',
+      loadFailed: '加载用户活动失败',
+      loadMore: '加载更多',
+      loading: '加载中...',
+      timeAgo: {
+        justNow: '刚刚',
+        minutesAgo: '{{count}}分钟前',
+        hoursAgo: '{{count}}小时前',
+        daysAgo: '{{count}}天前',
+        monthsAgo: '{{count}}个月前',
+        yearsAgo: '{{count}}年前',
+      },
+      types: {
+        rank: {
+          prefix: '在',
+          middle: '中获得了',
+          grade: '评级',
+          rankPrefix: '排名第',
+        },
+        rankLost: {
+          prefix: '在',
+          suffix: '中失去了第一名',
+        },
+        achievement: '获得了成就',
+        beatmapUpload: '上传了谱面',
+        beatmapUpdate: '更新了谱面',
+        beatmapRanked: '谱面被ranked',
+        beatmapLoved: '谱面获得Loved',
+        beatmapRevived: '谱面被复活',
+        beatmapDeleted: '谱面被删除',
+        usernameChange: '更改了用户名',
+        supportAgain: '再次成为了Supporter',
+        supportFirst: '首次成为了Supporter',
+        supportGift: '获得了Supporter赠送',
+        activity: '进行了活动',
+      },
+    },
+    bestScores: {
+      title: '最佳成绩',
+      noScores: '暂无最佳成绩',
+      loadFailed: '加载最佳成绩失败',
+      loadMore: '加载更多',
+      loading: '加载中...',
+      by: 'by',
+    },
+    friends: {
+      loading: '加载中...',
+    },
+    userPage: {
+      title: '个人介绍',
+      noContent: '还没有写个人介绍',
+      writeButton: '编写个人介绍',
+      editButton: '编辑',
+      editTitle: '编辑个人介绍',
+      placeholder: '使用 BBCode 编写你的个人介绍...',
+      editorPlaceholder: '为 {{username}} 编写个人介绍...\n\n你可以使用BBCode格式化文本，比如：\n[b]粗体文本[/b]\n[i]斜体文本[/i]\n[color=red]彩色文本[/color]\n\n点击工具栏按钮或使用快捷键来快速插入格式。',
+      processing: '内容正在处理中...',
+      cancel: '取消',
+      save: '保存',
+      saving: '保存中...',
+      saveSuccess: '个人页面已保存成功！',
+      saveError: '保存失败，请重试',
+      hasUnsavedChanges: '有未保存的更改',
+      confirmDiscard: '您有未保存的更改，确定要放弃编辑吗？',
+      noEditPermission: '您没有权限编辑此用户的页面',
+      cannotEditPage: '无法编辑此页面',
+      loadingEditor: '加载编辑器中...',
+      editPageTitle: '编辑个人页面',
+      editPageSubtitle: '为 {{username}} 编辑个人介绍',
+      unsavedChanges: '有未保存的更改',
+      cancelEditTooltip: '取消编辑 (Esc)',
+      saveShortcut: 'Ctrl+S 保存',
+      supportsBBCode: '支持BBCode格式',
+      bbcodeDescription: 'BBCode 是一种轻量级标记语言，用于格式化文本。您可以使用以下标签来美化您的个人介绍：',
+      usePreview: '使用预览功能查看最终效果',
+      editCover: '编辑封面',
+    },
+    friendActions: {
+      selfProfile: '自己的资料',
+      blocked: '已屏蔽该用户',
+      mutualFollow: '互相关注 - 你们相互关注',
+      following: '已关注 - 你关注了此用户',
+      followsYou: '关注你的用户 - 对方关注了你',
+      notFollowing: '未关注 - 点击查看关注选项',
+      unblock: '取消屏蔽',
+      cancelMutualFollow: '取消互相关注',
+      unfollow: '取消关注',
+      block: '屏蔽',
+      followBack: '回关 (互相关注)',
+      follow: '关注',
+    },
+    bbcodeEditor: {
+      placeholder: '在这里输入BBCode内容...',
+      toolbar: {
+        bold: '粗体',
+        italic: '斜体',
+        underline: '下划线',
+        strikethrough: '删除线',
+        spoiler: '剧透',
+        color: '颜色',
+        fontSize: '字体大小',
+        image: '插入图片',
+        link: '插入链接',
+        userProfile: '用户主页链接',
+        email: '邮箱链接',
+        youtube: 'YouTube视频',
+        audio: '音频',
+        imagemap: '图片映射',
+        quote: '引用',
+        code: '代码块',
+        list: '列表',
+        box: '折叠框',
+        center: '居中对齐',
+        heading: '标题',
+        notice: '通知框',
+        help: '帮助',
+      },
+      insertText: {
+        bold: '粗体文本',
+        italic: '斜体文本',
+        underline: '下划线文本',
+        strikethrough: '删除线文本',
+        colorText: '彩色文本',
+        text: '文本',
+        linkText: '链接文本',
+        username: '用户名',
+        emailLink: '邮箱链接',
+        clickToVisit: '点击访问网站',
+        infoArea: '这是信息区域',
+        quoteContent: '引用内容',
+        codeContent: '代码内容',
+        item1: '项目1',
+        item2: '项目2',
+        title: '标题',
+        collapsibleContent: '折叠内容',
+        spoilerContent: '剧透内容',
+        centerText: '居中文本',
+        headingText: '标题文本',
+        importantNotice: '重要通知',
+      },
+      colors: {
+        text: '色文本',
+      },
+      fontSize: {
+        label: '字体大小',
+        extraSmall: '极小',
+        small: '小',
+        normal: '普通',
+        large: '大',
+      },
+      help: {
+        title: 'BBCode帮助',
+        button: '帮助',
+        examples: {
+          bold: '粗体',
+          italic: '斜体',
+          underline: '下划线',
+          strikethrough: '删除线',
+          color: '彩色',
+          size: '大小',
+          link: '链接',
+          text: '文本',
+          imageUrl: '图片URL',
+          quote: '引用',
+          code: '代码',
+          item1: '项目1',
+          item2: '项目2',
+        },
+      },
+      modes: {
+        edit: '编辑',
+        preview: '预览',
+      },
+      preview: {
+        noContent: '输入内容以查看预览',
+        generateFailed: '无法生成预览',
+        generating: '生成预览中...',
+      },
+      validation: {
+        networkError: '验证失败，请检查网络连接',
+        validating: '验证中',
+        syntaxCorrect: '语法正确',
+        errors: '个错误',
+        syntaxErrors: 'BBCode语法错误',
+      },
+    },
+  },
+} as const;
