+++ src/i18n/locales/zh/pages/teams.ts
@@ -1,0 +1,124 @@
+export const teamsPage = {
+  teams: {
+    title: '战队排行榜',
+    description: '查看各个战队的表现和实力排名',
+    tabs: {
+      teams: '战队排行',
+    },
+    createTeam: '创建战队',
+    myTeam: '我的战队',
+    joinTeam: '加入战队',
+    viewTeam: '查看战队',
+    editTeam: '编辑战队',
+    teamRankings: '战队排行榜',
+    loadingTeams: '加载战队排行榜数据中...',
+    gameModes: {
+      osu: 'osu!',
+      taiko: 'osu!taiko',
+      fruits: 'osu!catch',
+      mania: 'osu!mania',
+    },
+    rankingTypes: {
+      performance: '表现分',
+      score: '分数',
+    },
+    rankingType: {
+      label: '排行类型',
+    },
+    subModes: {
+      vanilla: '常规',
+      relax: 'Relax',
+      autopilot: 'Autopilot',
+    },
+    teamCard: {
+      rank: '排名',
+      members: '成员',
+      captain: '队长',
+      averageRank: '平均排名',
+      totalPerformance: '总表现分',
+      country: '国家',
+    },
+    pagination: {
+      previous: '上一页',
+      next: '下一页',
+      page: '第 {{current}} 页 / 共 {{total}} 页',
+    },
+    errors: {
+      loadFailed: '加载战队失败',
+      noData: '暂无战队数据',
+    },
+    detail: {
+      createdAt: '创建于 {{date}}',
+      members: '{{count}} 名成员',
+      captain: '队长',
+      membersSectionTitle: '成员',
+      teamMembers: '团队成员',
+      memberCount: '({{count}} 人)',
+      loading: '加载战队详情中...',
+      notFound: '战队不存在',
+      notFoundDescription: '未找到该战队信息',
+      backToTeams: '返回战队列表',
+      teamInfo: '战队头部信息',
+      coverImage: '封面图片',
+      teamFlag: '战队旗帜',
+      teamBasicInfo: '战队基本信息',
+      teamActions: '战队操作按钮',
+      captainInfo: '队长信息',
+      memberActions: '成员操作按钮',
+      joinTeam: '请求加入',
+      joining: '请求中...',
+      leaveTeam: '退出战队',
+      leaving: '退出中...',
+      editTeam: '编辑战队',
+      deleteTeam: '删除战队',
+      kickMember: '踢出',
+      kicking: '踢出中...',
+      joinRequestSent: '加入请求已发送，请等待队长审核',
+      leftTeam: '已退出战队',
+      confirmLeave: '确定要退出这个战队吗？',
+      confirmDelete: '确定要删除这个战队吗？此操作不可撤销！',
+      confirmKick: '确定要踢出 {{username}} 吗？',
+      teamDeleted: '战队已删除',
+      memberKicked: '已将 {{username}} 踢出战队',
+    },
+    create: {
+      loginRequired: '您需要登录后才能创建或编辑战队',
+      nameRequired: '请输入战队名称',
+      shortNameRequired: '请输入战队简称',
+      assetsRequired: '请上传战队旗帜和封面',
+      updateSuccess: '战队信息更新成功',
+      createSuccess: '战队创建成功',
+      loading: '加载战队信息中...',
+      editTeam: '编辑战队',
+      createTeam: '创建战队',
+      editDescription: '修改您的战队信息',
+      createDescription: '创建一个新的战队，邀请朋友一起游戏',
+      basicInfo: '基本信息',
+      teamName: '战队名称',
+      teamNamePlaceholder: '输入战队名称',
+      teamShortName: '战队简称',
+      teamShortNamePlaceholder: '输入战队简称',
+      shortNameDescription: '简称将显示在排行榜中，建议使用 2-5 个字符',
+      teamFlag: '战队旗帜',
+      selectFlag: '选择旗帜',
+      flagDescription: '标准尺寸: 240×120px，最大 2MB，支持裁剪调整',
+      teamCover: '战队封面',
+      selectCover: '选择封面',
+      coverDescription: '推荐尺寸: 1920×1280px，最大 10MB，支持裁剪调整',
+      creatingTeam: '创建战队中...',
+      memberManagement: '队员管理',
+      leaderTransfer: '队长转让 (可选)',
+      keepCurrentLeader: '保持当前队长',
+      leaderTransferDescription: '选择一个新的队长，如果不选择则保持当前队长不变。转让队长权限后，您将失去管理权限。',
+      currentMembers: '当前队员 ({{count}} 人)',
+      currentLeader: '当前队长',
+      willBecomeLeader: '将成为新队长',
+      noMembersAvailable: '没有可选择的成员',
+      unknown: '未知',
+      cancel: '取消',
+      saving: '保存中...',
+      saveChanges: '保存修改',
+      required: '*',
+    },
+  },
+} as const;
